+++ sandbox/stashbox.benchmarks/Program.cs
@@ -11,7 +11,7 @@ namespace Stashbox.Benchmarks
             //BenchmarkRunner.Run<EnumerableBenchmarks>();
             //BenchmarkRunner.Run<PropertyBenchmarks>();
             //BenchmarkRunner.Run<RegisterBenchmarks>();
-            BenchmarkRunner.Run<ResolveBenchmarks>();
+            //BenchmarkRunner.Run<ResolveBenchmarks>();
             //BenchmarkRunner.Run<ScopedBenchmarks>();
             //BenchmarkRunner.Run<FuncBenchmarks>();
             //BenchmarkRunner.Run<FinalizerBenchmarks>();