+++ src/Lifetime/ScopedLifetime.cs
@@ -32,7 +32,7 @@ namespace Stashbox.Lifetime
                     serviceRegistration.RegistrationId.AsConstant(),
                     factory.AsConstant(),
                     resolutionContext.RequestContextParameter,
-                    resolveType.AsConstant())
+                    serviceRegistration.ImplementationType.AsConstant())
                 .ConvertTo(resolveType);
         }
     }