+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v5.0.1] - 2022-02-10
+### Changed
+- Converted the `ServiceContext` to a read-only struct.
+- Made the `.AsServiceContext()` extension method of `Expression` public.
+
 ## [v5.0.0] - 2022-02-09
 ### Added
 - Additional [metadata registration](https://z4kn4fein.github.io/stashbox/#/advanced/generics?id=metadata-amp-tuple) option.
@@ -192,6 +197,7 @@ The validation was executed only at the expression tree building phase, so an al
 - Removed the legacy container extension functionality.
 - Removed the support of PCL v259.
 
+[v5.0.1]: https://github.com/z4kn4fein/stashbox/compare/5.0.0...5.0.1
 [v5.0.0]: https://github.com/z4kn4fein/stashbox/compare/4.1.0...5.0.0
 [v4.1.0]: https://github.com/z4kn4fein/stashbox/compare/4.0.0...4.1.0
 [v4.0.0]: https://github.com/z4kn4fein/stashbox/compare/3.6.4...4.0.0